requirements.txt
flask
websocket-client
yaml
pillow